+++ CHANGELOG.md
@@ -9,5 +9,5 @@
 
 ### Additions
 - Added `const_nzbw` and `const_bw` to `InlAwi`
-- Added `lut_set`, `mul_assign`, and `arb_umul_add_assign` to `Bits`
+- Added `lut_set`, `mul_assign`, `arb_umul_add_assign`, and `arb_imul_add_assign` to `Bits`
 - Added `from_bytes_general` to `ExtAwi`